+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- fix: focusElm accessing `document` at initialization
 
 ## 4.8.0 – 4.8.1